backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
kiteconnect==5.0.1
python-dotenv==1.0.0
websockets==12.0
cryptography>=41.0.0
